Fix type truncation warning that breaks build on Windows
Inside `UnsignedUnitTest`, we are passing `unit32_t` into the `shift` argument
of several opcodes (e.g., `OpCode::Rol`), which are `unit8_t`. Since we
interpret warnings as errors, this breaks the build on Windows.

This commit will change the argument type to be `uint8_t`, which fixes the
Windows build.

+++ UnitTest/NativeJITUnitTest/UnsignedUnitTest.cpp
@@ -232,7 +232,7 @@ namespace NativeJIT
             {
                 Function<uint32_t, uint32_t, uint32_t> expression(setup->GetAllocator(), setup->GetCode());
 
-                uint32_t p2 = 1ul;
+                uint8_t p2 = 1ul;
 
                 auto & a = expression.Rol(expression.GetP1(), p2);
                 auto function = expression.Compile(a);
@@ -273,7 +273,7 @@ namespace NativeJIT
             {
                 Function<uint32_t, uint32_t, uint32_t> expression(setup->GetAllocator(), setup->GetCode());
 
-                uint32_t p2 = 3ul;
+                uint8_t p2 = 3ul;
 
                 auto & a = expression.Shl(expression.GetP1(), p2);
                 auto function = expression.Compile(a);
@@ -294,7 +294,7 @@ namespace NativeJIT
             {
                 Function<uint32_t, uint32_t, uint32_t> expression(setup->GetAllocator(), setup->GetCode());
 
-                uint32_t p2 = 3ul;
+                uint8_t p2 = 3ul;
 
                 auto & a = expression.Shr(expression.GetP1(), p2);
                 auto function = expression.Compile(a);
